+++ models/DefaultPromptInput.ts
@@ -16,6 +16,8 @@ class DefaultPromptInput {
   loras: Lora[] = []
   maskData: any | null = ''
   models: Array<string> = ['stable_diffusion']
+  multiClip: string = ''
+  multiDenoise: string = ''
   multiGuidance: string = ''
   multiSteps: string = ''
   negative: string = ''
@@ -37,6 +39,8 @@ class DefaultPromptInput {
   useAllModels: boolean = false
   useAllSamplers: boolean = false
   useFavoriteModels: boolean = false
+  useMultiClip: boolean = false
+  useMultiDenoise: boolean = false
   useMultiGuidance: boolean = false
   useMultiSteps: boolean = false
   width: number = 512